% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_torsten.R
\name{install_torsten}
\alias{install_torsten}
\title{Installation for torsten}
\usage{
install_torsten(StanHeaders_version = NULL, rstan_version = NULL,
  lib = .libPaths()[1], ...)
}
\arguments{
\item{StanHeaders_version}{package_version, package version of StanHeaders to append Torsten Default: NULL}

\item{rstan_version}{package_version, package version of rstan to install, Default: NULL}

\item{lib}{character, giving the library directory where to install the packages, Default: .libPaths()[1]}

\item{...}{parameters to pass to install.packages}
}
\description{
installation of torsten with rstan, that replaces stanHeaders.
}
\details{
installation will replace the 'StanHeaders/include/src/stan' and 'StanHeaders/include/stan' of stanHeaders and install
source rstan without dependencies.
}
