% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_torsten_remote.R
\name{install_torsten_remote}
\alias{install_torsten_remote}
\title{Installation for torsten through remote repository}
\usage{
install_torsten_remote(StanHeaders_version = NULL, rstan_version = NULL,
  math_branch = "torsten-master", stan_branch = "torsten-master",
  lib = .libPaths()[1], ...)
}
\arguments{
\item{StanHeaders_version}{package_version, package version of StanHeaders to append Torsten Default: NULL}

\item{rstan_version}{package_version, package version of rstan to install, Default: NULL}

\item{math_branch}{character, install the current build ('master') or different
development branch of torsten math library (e.g. 'develop'), Default: 'torsten-master'}

\item{stan_branch}{character, install the current build ('master') or different
development branch of torsten stan library (e.g. 'develop'), Default: 'torsten-master'}

\item{lib}{character, giving the library directory where to install the packages, Default: .libPaths()[1]}

\item{...}{parameters to pass to install.packages}
}
\description{
installation of torsten with rstan, that replaces stanHeaders.
This function allows to specify which branch of torsten to install from.
}
\details{
installation will replace the 'StanHeaders/include/src/stan' and 'StanHeaders/include/stan' of stanHeaders and install
source rstan without dependencies.
}
